connected leaderboard to flask

+++ web-socket/websock.py
@@ -10,16 +10,6 @@ dir_path = os.path.dirname(os.path.realpath(__file__))
 app = Flask(__name__, static_folder='../web-ui/build', static_url_path='/')
 sock = Sock(app)
 count = 0
-scores = [
-    {"person": "Clay", "value": 1},
-    {"person": "Tyler", "value": 2},
-    {"person": "Sam", "value": 3},
-    {"person": "Adhishri", "value": 4},
-    {"person": "Anmol", "value": 5},
-    {"person": "Siddharth", "value": 6},
-    {"person": "Ajay", "value": 7},
-    {"person": "Vijay", "value": 8},
-]
 
 
 def update_score_message(sock, num):
@@ -27,8 +17,15 @@ def update_score_message(sock, num):
         'action': 'update_score',
         'score': num,
     })
-    print(f"<-- {msg}")
+    # print(f"<-- {msg}")
     sock.send(msg)
+
+
+def scores_to_json(cursor):
+    cursor.execute("select * from scores order by score desc")
+    scores = cursor.fetchall()
+    lst = [{'person': name, 'value': score} for name, score in scores]
+    return lst
 
 
 def update_leaderboard_message(sock, ldb):
@@ -36,35 +33,38 @@ def update_leaderboard_message(sock, ldb):
         'action': 'update_leaderboard',
         'leaderboard': ldb,
     })
-    print(f"<-- {msg}")
     sock.send(msg)
+
+
+def receive_ws(sock, cursor, connection):
+    global count
+    msg = sock.receive()
+    if msg is not None:
+        msg = json.loads(msg)
+        if msg['action'] == 'start_game':
+            args = (msg['currentPlayer'], 0, count, count)
+            cursor.execute(
+                f"insert into scores (person, score) values (?, ?) ON CONFLICT (person) DO UPDATE SET score = ? WHERE ? > score", args)
+            connection.commit()
+        elif msg['action'] == 'reset_leaderboard':
+            cursor.execute("delete from scores where score >= 0")
+            connection.commit()
+        else:
+            print("unknown action")
 
 
 @sock.route('/socket')
 def echo(sock):
-    print(dir_path)
     connection = sqlite3.connect(f'{dir_path}/ScoreDB/scores.db')
     cursor = connection.cursor()
     cursor.execute(
         "create table if not exists scores (person text PRIMARY KEY, score integer)")
-    cursor.execute("delete from scores where score >=0")
     while True:
         global count
-        if count < len(scores):
-            print("did")
-            args = (scores[count]['person'], scores[count]['value'], scores[count]['value'], scores[count]['value'])
-            cursor.execute(
-                f"insert into scores (person, score) values (?, ?) ON CONFLICT (person) DO UPDATE SET score = ? WHERE ? > score", args)
-            connection.commit()
-            update_leaderboard_message(sock, scores[:count])
-        elif count == len(scores):
-            cursor.execute(
-                "select * from scores")
-            search = cursor.fetchall()
-            print(search)
-            connection.close()
+        update_leaderboard_message(sock, scores_to_json(cursor))
         count += 1
         update_score_message(sock, count)
+        receive_ws(sock, cursor, connection)
         sleep(1)
         # score = SerialRead()["sensorVoltage"]
